feat: allow partial name search in monster API
Updated the API to support searching for monsters based on partial name
matches using regular expressions. Now users can search for monsters by
entering a substring in the search bar, making it more user-friendly.
Also added the ability to trigger the search by pressing the "Enter" key
in the search input field on the frontend.